# hr_website/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HR Website API",
    "version": "1.0.0",
    "description": "API documentation for the HR Website backend."
  },
  "components": {
    "schemas": {
...
      "Branch": {
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "manager_id",
          "latitude",
          "longitude",
          "attendance_radius",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "manager_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Chi nhánh trưởng (public.users.id)."
          },
          "latitude": {
            "type": "number",
            "nullable": true,
            "minimum": 8.5436,
            "maximum": 23.3883,
            "description": "Vĩ độ, chỉ chấp nhận toạ độ tại Việt Nam (8.5436–23.3883° N)."
          },
          "longitude": {
            "type": "number",
            "nullable": true,
            "minimum": 102.0967,
            "maximum": 109.4944,
            "description": "Kinh độ, chỉ chấp nhận toạ độ tại Việt Nam (102.0967–109.4944° E)."
          },
          "attendance_radius": {
            "type": "integer",
            "description": "Bán kính cho phép chấm công (mét) quanh toạ độ chi nhánh."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "EmployeeStatus": {
        "type": "object",
        "required": [
          "work_date",
          "employee_id",
          "full_name",
          "role",
          "branch_id",
          "branch_name",
          "status",
          "shift_name",
          "check_in_at",
          "check_out_at",
          "check_in_distance_m",
          "attendance_radius"
        ],
        "properties": {
          "work_date": {
            "type": "string",
            "format": "date",
            "description": "Ngày nghiệp vụ (Asia/Bangkok) mà bản ghi trạng thái mô tả."
          },
          "employee_id": {
            "type": "string",
            "format": "uuid"
          },
          "full_name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MANAGER",
              "EMPLOYEE"
            ],
            "description": "Vai trò hệ thống: OWNER, MANAGER hoặc EMPLOYEE (SCRUM-59)."
          },
          "branch_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "branch_name": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "WORKING",
              "NOT_STARTED",
              "ON_LEAVE",
              "FINISHED",
              "NO_SHIFT",
              "RESIGNED"
            ],
            "description": "Đang làm việc, chưa vào ca, nghỉ phép, đã tan ca, không có ca hôm nay, đã nghỉ việc."
          },
          "shift_name": {
            "type": "string",
            "nullable": true
          },
          "check_in_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "check_out_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "check_in_distance_m": {
            "type": "integer",
            "nullable": true
          },
          "attendance_radius": {
            "type": "integer",
            "nullable": true,
            "description": "Bán kính cho phép của chi nhánh, để so với khoảng cách chấm công."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Mô tả lỗi.",
            "example": "Trường `question` là bắt buộc và phải là chuỗi không rỗng."
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/employee-status": {
      "get": {
        "summary": "Get the live employee status of one business day",
        "description": "One row per person for the day (SCRUM-22), derived from `shift_assignments` and\n`attendance` by the `get_employee_status` RPC: đang làm việc, chưa vào ca, nghỉ phép,\nđã tan ca, không có ca hôm nay or nghỉ việc. The day defaults to today in Asia/Bangkok.\nSCRUM-59: giám sát nhân sự is a management view (`employee-status:view`), so a plain\nemployee gets `403`; SCRUM-60 narrows the rows `get_employee_status` may return.\n",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branch_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Only the people scheduled to, or checking in at, this branch."
          },
          {
            "in": "query",
            "name": "work_date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Business day to report (defaults to today in Asia/Bangkok)."
          }
        ],
        "responses": {
          "200": {
            "description": "The snapshot and the day it describes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "updated_at",
                    "work_date",
                    "data"
                  ],
                  "properties": {
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "work_date": {
                      "type": "string",
                      "format": "date"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmployeeStatus"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "branch_id is not a UUID, or work_date is not a calendar date.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller has no user profile, or their role may not read the roster (SCRUM-59).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/branches": {
      "get": {
        "summary": "List branches",
        "description": "Returns every branch visible to the authenticated caller under the active RLS policy, sorted by name.",
        "tags": [
          "Branches"
        ],
        "responses": {
          "200": {
            "description": "The branch list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "500": {
            "description": "An unexpected server error occurred."
          }
        }
      },
      "post": {
        "summary": "Create a branch",
        "description": "Creates a branch together with its GPS geofence. Restricted to the organization's OWNER (`branch:manage`, SCRUM-59) and enforced by RLS.",
        "tags": [
          "Branches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 120,
                    "example": "Chi nhánh Quận 1 - Flagship Bến Thành"
                  },
                  "address": {
                    "type": "string",
                    "maxLength": 300,
                    "nullable": true,
                    "example": "128 Lê Lợi, Phường Bến Thành, Quận 1, TP. Hồ Chí Minh"
                  },
                  "manager_id": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true,
                    "description": "Chi nhánh trưởng — `public.users.id` của một tài khoản MANAGER trong tổ chức của bạn (SCRUM-63: chủ sở hữu đứng trên mọi chi nhánh nên không phụ trách một chi nhánh cụ thể); bỏ trống hoặc null là chưa gán. `400` khi id không trỏ tới tài khoản nào mà bạn nhìn thấy được, hoặc trỏ tới một nhân viên hay chủ sở hữu."
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": 8.5436,
                    "maximum": 23.3883,
                    "nullable": true,
                    "example": 10.7725,
                    "description": "Vĩ độ; phải nằm trên lãnh thổ Việt Nam — kiểm tra theo đường bờ biển thật (lib/geo/vietnamOutline.ts), không chỉ theo khung 8.5436–23.3883° N."
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": 102.0967,
                    "maximum": 109.4944,
                    "nullable": true,
                    "example": 106.698,
                    "description": "Kinh độ; phải nằm trên lãnh thổ Việt Nam — kiểm tra theo đường bờ biển thật (lib/geo/vietnamOutline.ts), không chỉ theo khung 102.0967–109.4944° E."
                  },
                  "attendance_radius": {
                    "type": "integer",
                    "minimum": 10,
                    "maximum": 2000,
                    "default": 50,
                    "description": "Allowed check-in distance in metres (defaults to 50)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The branch was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid, including a manager_id that names no account in your organization or names an account that is not a MANAGER (SCRUM-63)."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The caller's role may not manage branches (SCRUM-59, owner only)."
          },
          "500": {
            "description": "An unexpected server error occurred."
          }
        }
      }
    },
...
  },
...
}